Introduces our own custom commit module loader called (uncreatively) require. The module loader is designed to mostly follow the AMDJS spec with the additional rule that if a module starts with local:// it will "jump the bridge". It does not do everything, most things missing are pointed out in the implementation.
A custom module loader was needed because none of those tested fit all
of our needs while being small enough. Not to mention none could handle
loading from across the bridge (although using AMDJS plugin loaders may
have worked).

This change also includes changes to the server side
(bootstrap/index.html) because AMDJS does not have a "load" method.
Instead we use the require(dependencies, Function) method. This
required changes to jnext.js to auto bind to window.

This also includes changes to the framework to remove the assumption of
using window.require since on the client side we now closure require so
it does not leak to the global namespace. This also allows us to remove
all logic that was used to distinguish between module loaders
(requireLocal) except in the builder where we use the local:// protocol.

Also includes changes to jshint ignore and setup to allow lint to pass. I
had to turn off the node settings which meant I needed to add some
predefines. I think this is a better approach anyway since we are only
using node for testing.