Updating the webworks framework to remove all reliance on the lower level qnx.callExtension calls and begin using the webplatform.js. This should add a point of stability that we can maintain across webworks revisions.
This relies on the webplatform, which we have added as a submodule. This
means before the WebWorks framework can be built, the webplatform
framework must be built. To accomodate this, configure script changes
were made. These changes will build the submodule everywhere, but WILL
NOT check them out on SCM.

Also includes framework cleanup to reduce and centralize all the
confusing request logic into a single module with an understandable
workflow.

Also includes unit tests for this work which were previously missing
from the framework. As well as minor fixes for the dialog tests to get
them passing.