This commit enhances the automated test suite available for testing webworks whitelisting.
It includes changes to the domains used by the test-app for the
whitelisting tests and updates the config.xml to have it pass.

It also includes several addional apps to test other aspects of
whitelisting that could not be done from a single app.

Additionally it includes some unit test additions to make sure we have
complete coverage.

Reviewed By: Jeffrey Heifetz <jheifetz@rim.com>
Tested By: Igor Shneur <ishneur@rim.com>