Updated test-app to run in Ripple
Modified the way specs are pulled in to use relative paths
rather then the local:/// scheme which will not work in
ripple

Also fixed a small issue with config.xml. It was being interpreted
as being malformed in the Chrome browser.

Tested in both Ripple and on device to ensure path changes work.