Adds the ability run a test-suite.
Adds the ability to run a test suite using jasmine in node to control
deployment and termination of apps. It uses a child node server to
get the results back from the app that is currently running and displays
the results on the console on desktop.

Reorganizes the test apps to be located under the test-suite folder.

To run these tests you must use
    jake test-suite
The options are controlled under a test-runner.json file which holds all
the properties for runnning tests like packager location, and device ip
and password.

To run one a particular part of the suite use -

    jake test-suite test/test-suite/test/whitelist-suite-runner.js

Only whitelist tests have been added to the test suite for now, other
will moved slowly along with the main test-app as well.