Added description to invalid argument error when saving a contact
Ran into this when developing cordova's contact plugin. I would get an
error about an invalid argument but no indication of what needed to be
fixed (or how). Since the validation requirements on the contact object
are not documented anywhere I had to go into the code to find them out.

There is now a message on the exception that will guide the developer to
know what is incorrect on the contact they are attempting to save

Reviewed By: Rosa Tse
Tested By: Rosa Tse